+++ spec/services/webdav_data_file_synchronizer_spec.rb
@@ -59,7 +59,7 @@ describe WebdavDataFileSynchronizer, files: true do
         expect { call(test_patient, correct_user) }.to change { DataFile.count }.by(2)
         expect(DataFile.all.map(&:data_type)).
           to match_array %w[fluid_virtual_model ventricle_virtual_model]
-        expect(DataFile.all.map(&:pipeline_id).compact).to be_empty
+        expect(DataFile.all.map(&:output_of_id).compact).to be_empty
       end
 
       it 'only creates input data_files not yet present' do
@@ -69,7 +69,7 @@ describe WebdavDataFileSynchronizer, files: true do
         expect { call(test_patient, correct_user) }.to change { DataFile.count }.by(1)
         expect(DataFile.all.map(&:data_type)).
           to match_array %w[fluid_virtual_model ventricle_virtual_model]
-        expect(DataFile.all.map(&:pipeline_id).compact).to be_empty
+        expect(DataFile.all.map(&:output_of_id).compact).to be_empty
       end
 
       it 'recognizes files with regexps' do
@@ -97,31 +97,43 @@ describe WebdavDataFileSynchronizer, files: true do
 
     context 'for a given pipeline' do
       let(:test_patient_with_pipeline) { create(:patient, :with_pipeline) }
+      let(:pipeline) { test_patient_with_pipeline.pipelines.first }
 
       it 'calls file storage and creates new pipeline-related data_files' do
         expect { call(test_patient_with_pipeline, correct_user) }.
-          to change { DataFile.count }.by(1)
-        expect(DataFile.first.data_type).to eq 'blood_flow_result'
-        expect(DataFile.first.pipeline).to eq test_patient_with_pipeline.pipelines.first
+          to change { DataFile.count }.by(2)
+        expect(pipeline.inputs.first.data_type).to eq 'ventricle_virtual_model'
+        expect(pipeline.outputs.first.data_type).to eq 'blood_flow_result'
       end
 
       it 'only creates pipeline data_files not yet present' do
+        create(:data_file, name: 'structural_vent.dat',
+                           data_type: 'ventricle_virtual_model',
+                           patient: test_patient_with_pipeline,
+                           input_of: pipeline)
         create(:data_file, name: 'fluidFlow-1-00002.dat',
                            data_type: 'blood_flow_result',
                            patient: test_patient_with_pipeline,
-                           pipeline: test_patient_with_pipeline.pipelines.first)
+                           output_of: pipeline)
+
         expect { call(test_patient_with_pipeline, correct_user) }.to change { DataFile.count }.by(0)
       end
 
       it 'destroys pipeline data_files which are no longer stored in File Storage' do
-        create(:data_file, name: 'structural_vent.dat',
+        create(:data_file, name: 'structural_vent1.dat',
                            data_type: 'ventricle_virtual_model',
                            patient: test_patient_with_pipeline,
-                           pipeline: test_patient_with_pipeline.pipelines.first)
+                           input_of: pipeline)
+        create(:data_file, name: 'structural_vent2.dat',
+                           data_type: 'ventricle_virtual_model',
+                           patient: test_patient_with_pipeline,
+                           output_of: pipeline)
+
         call(test_patient_with_pipeline, correct_user)
-        expect(DataFile.count).to eq 1
-        expect(DataFile.first.data_type).to eq 'blood_flow_result'
-        expect(DataFile.first.pipeline).to eq test_patient_with_pipeline.pipelines.first
+
+        expect(DataFile.count).to eq 2
+        expect(pipeline.inputs.first.data_type).to eq 'ventricle_virtual_model'
+        expect(pipeline.outputs.first.data_type).to eq 'blood_flow_result'
       end
     end
   end